=Update add_rewrite_endpoint to support endpoint only rules.
Increased popularity of the amp plugin has shown that matching all elements that come after an endpoint can lead to the desired page being accessible across invalid urls. By enabling a restriction so that the rewrite rule only matches the endpoint /?$ we ensure that urls go to 404 responses vs 200 and leading to indexing of bad links across websites. Reducing the match of endpoints also can lead to cleaner rewrites.